Move classes from navigation.navigation to where they fit
