+++ appyx-navigation/common/src/commonMain/kotlin/com/bumble/appyx/navigation/children/ChildNodeBuilder.kt
@@ -1,4 +1,4 @@
-package com.bumble.appyx.navigation.navigation
+package com.bumble.appyx.navigation.children
 
 import com.bumble.appyx.navigation.modality.BuildContext
 import com.bumble.appyx.navigation.node.Node
