+++ appyx-navigation/common/src/commonMain/kotlin/com/bumble/appyx/navigation/integration/UpNavigationHandler.kt
@@ -1,4 +1,4 @@
-package com.bumble.appyx.navigation.navigation.upnavigation
+package com.bumble.appyx.navigation.integration
 
 fun interface UpNavigationHandler {
     fun handleUpNavigation()
