+++ documentation/releases/2.0.0-alpha10.md
@@ -6,7 +6,15 @@ title: 2.0.0-alpha10 – Migration guide
 
 ## Package changes
 
-- Merged `com.bumble.appyx.navigation.integrationpoint` into `com.bumble.appyx.navigation.integration`
+```diff
+// Moved all classes:
+-com.bumble.appyx.navigation.integrationpoint.*
++com.bumble.appyx.navigation.integration.*
+
+// Moved class:
+-com.bumble.appyx.navigation.navigation.UpNavigationHandler
++com.bumble.appyx.navigation.integration.UpNavigationHandler
+```
 
 ## Rename ambiguous `AppyxComponent` composables
